requirements.txt
aniso8601==9.0.1
asgiref==3.5.2
async-timeout==4.0.2
certifi==2022.12.7
cffi==1.15.1
channels==4.0.0
channels-redis==4.0.0
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==39.0.0
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.1.4
django-author==1.2.0
django-birthday==0.1.4
django-bookmark-base==2020.7.1
django-cors-headers==3.13.0
django-crispy-forms==1.14.0
django-environ==0.9.0
django-filter==22.1
django-import-export==3.0.1
django-js-asset==2.0.0
django-mptt==0.14.0
django-phone-field==1.8.1
django-tables2==2.4.1
django-templated-mail==1.1.1
django-tinymce==3.5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-yasg==1.21.4
et-xmlfile==1.1.0
graphene==3.1.1
graphene-django==3.0.0
graphql-core==3.2.3
graphql-relay==3.2.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
Markdown==3.4.1
MarkupPy==1.14
MarkupSafe==2.1.1
msgpack==1.0.4
oauthlib==3.2.2
odfpy==1.4.1
openpyxl==3.0.10
packaging==22.0
Pillow==9.3.0
promise==2.3
pycparser==2.21
PyJWT==2.6.0
python-barcode==0.14.0
python3-openid==3.2.0
pytz==2022.6
PyYAML==6.0
redis==4.4.1
requests==2.28.1
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
setuptools-git==1.2
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
tablib==3.2.1
text-unidecode==1.3
uritemplate==4.1.1
urllib3==1.26.13
xlrd==2.0.1
xlwt==1.3.0
